Move all casters to pybind.h

[ghstack-poisoned]
