+++ torch/csrc/utils/python_arg_parser.h
@@ -79,82 +79,6 @@
 #include <string>
 #include <vector>
 
-namespace pybind11 {
-namespace detail {
-template <>
-struct type_caster<c10::SymInt> {
- public:
-  PYBIND11_TYPE_CASTER(c10::SymInt, _("SymInt"));
-  bool load(py::handle src, bool) {
-    if (torch::is_symint(src)) {
-      value = c10::SymInt(static_cast<c10::SymNode>(
-          c10::make_intrusive<torch::impl::PythonSymNodeImpl>(
-              src.attr("node"))));
-      return true;
-    }
-
-    auto raw_obj = src.ptr();
-    if (THPUtils_checkIndex(raw_obj)) {
-      value = c10::SymInt{THPUtils_unpackIndex(raw_obj)};
-      return true;
-    }
-    return false;
-  }
-
-  static py::handle cast(
-      c10::SymInt si,
-      return_value_policy /* policy */,
-      handle /* parent */) {
-    if (si.is_symbolic()) {
-      // TODO: generalize this to work with C++ backed class
-      auto* py_node = dynamic_cast<torch::impl::PythonSymNodeImpl*>(
-          si.toSymNodeImpl().get());
-      TORCH_INTERNAL_ASSERT(py_node);
-      return torch::get_symint_class()(py_node->getPyObj()).release();
-    } else {
-      return py::cast(si.as_int_unchecked()).release();
-    }
-  }
-};
-
-template <>
-struct type_caster<c10::SymFloat> {
- public:
-  PYBIND11_TYPE_CASTER(c10::SymFloat, _("SymFloat"));
-  bool load(py::handle src, bool) {
-    if (torch::is_symfloat(src)) {
-      value = c10::SymFloat(static_cast<c10::SymNode>(
-          c10::make_intrusive<torch::impl::PythonSymNodeImpl>(
-              src.attr("node"))));
-      return true;
-    }
-
-    auto raw_obj = src.ptr();
-    if (THPUtils_checkDouble(raw_obj)) {
-      value = c10::SymFloat{THPUtils_unpackDouble(raw_obj)};
-      return true;
-    }
-    return false;
-  }
-
-  static py::handle cast(
-      c10::SymFloat si,
-      return_value_policy /* policy */,
-      handle /* parent */) {
-    if (si.is_symbolic()) {
-      // TODO: generalize this to work with C++ backed class
-      auto* py_node = dynamic_cast<torch::impl::PythonSymNodeImpl*>(
-          si.toSymNodeImpl().get());
-      TORCH_INTERNAL_ASSERT(py_node);
-      return torch::get_symfloat_class()(py_node->getPyObj()).release();
-    } else {
-      return py::cast(si.as_float_unchecked()).release();
-    }
-  }
-};
-} // namespace detail
-} // namespace pybind11
-
 inline bool THPUtils_checkScalar(PyObject* obj) {
 #ifdef USE_NUMPY
   if (torch::utils::is_numpy_scalar(obj)) {
